# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel API",
        "description": "Sayohat veb-sayti uchun API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://toqtarbay.dbc.uz/api",
            "description": "Production Server"
        },
        {
            "url": "http://travel.test/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/contact": {
            "get": {
                "tags": [
                    "Contact"
                ],
                "summary": "Bog'lanish ma'lumotlarini olish",
                "operationId": "fc65a9ced8387969bc5b37c8eeafb9db",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Til kodi (languages jadvalidagi tillar)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "uz"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muvaffaqiyatli javob",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "+998901234567"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "info@example.com"
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "Tashkent, Uzbekistan"
                                                },
                                                "longitude": {
                                                    "type": "string",
                                                    "example": "69.2401"
                                                },
                                                "latitude": {
                                                    "type": "string",
                                                    "example": "41.2995"
                                                },
                                                "telegram_url": {
                                                    "type": "string",
                                                    "example": "https://t.me/example"
                                                },
                                                "telegram_username": {
                                                    "type": "string",
                                                    "example": "@example"
                                                },
                                                "instagram_url": {
                                                    "type": "string",
                                                    "example": "https://instagram.com/example"
                                                },
                                                "facebook_url": {
                                                    "type": "string",
                                                    "example": "https://facebook.com/example"
                                                },
                                                "facebook_name": {
                                                    "type": "string",
                                                    "example": "Example Page"
                                                },
                                                "youtube_url": {
                                                    "type": "string",
                                                    "example": "https://youtube.com/example"
                                                },
                                                "whatsapp_phone": {
                                                    "type": "string",
                                                    "example": "+998901234567"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ma'lumot topilmadi"
                    }
                }
            }
        },
        "/contact/banner": {
            "get": {
                "tags": [
                    "Contact"
                ],
                "summary": "Contact sahifasi bannerini olish",
                "operationId": "e9633877715fe78c912e9bfde2b84ec6",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Til kodi (kk, uz, ru, en)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "uz"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muvaffaqiyatli javob",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Biz bilan bog'laning"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "/storage/uploads/contact-banner.jpg"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Banner topilmadi"
                    }
                }
            }
        },
...
    },
...
}